refactor(ledger): extract transcript conformance tests into a dedicated module
Move the transcript and paging conformance tests from the monolithic conformance.rs file into a separate transcripts.rs module, reducing the main file by over 170 lines and improving maintainability by grouping related test helpers together.

Auto-committed-on: dragonfly
